Add GUI for Analyzer
Using PyInstaller package, developer can generate a simple GUI for the
network analyzer. The GUI gives the user the ability to choose the relay
and the rpc servers for generating the data, and the folder at which the
csv data files are going to be generated.
For building the app, just run `pyinstaller app.py -y`, this will
generate the executable and other files at `./dist/app`, and then
you will need to maually copy `contracts.json` to
`./dist/app/network_analyzer/contracts.json`.
Worths mentioning that `PyInstaller` generates the executable for
the environment it runs on (generates .exex on a Windows machine
and mac executable on Mac). Also, there are limitations on Mac
machine for the specific OS version. PyInstaller is stupid enough to
force you to make a run on each MacOS version you want the app to run on